requirements.txt
pyautogui
pyyaml
pillow
opencv-python
pytesseract
pygetwindow
selenium
playwright
keyboard
pyperclip